Some performance improvements

+++ src/headers/neurons_cpu.h
@@ -3,7 +3,7 @@
 
 #include <stdlib.h>
 
-extern double ***cpu_feedforward(const size_t rows, const size_t columns_Y, const size_t columns_X, const int layers,
+extern double ***cpu_feedforward(const int rows, const int columns_Y, const int columns_X, const int layers,
                                  double *X, double ***wb,
                                  const int nodes[layers], char funcs[layers+1][30]);
