+++ src/utils/loss.c
@@ -1,8 +1,6 @@
-#include <stdio.h>
-#include <stdlib.h>
 #include <math.h>
 
-double rmse(const size_t rows, const size_t columns_Y, const double *Y, const double *prediction) {
+double rmse(const int rows, const int columns_Y, const double *Y, const double *prediction) {
   int i, thresh = rows * columns_Y;
   double loss = 0.0;
   double dif;
