+++ src/utils/randomization.c
@@ -1,4 +1,3 @@
-#include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
 
@@ -11,7 +10,7 @@ void swap(double *a, double *b)
 }
 
 // A function to generate a random permutation of inputs
-void randomize(double *X, const size_t rows, const size_t columns_X)
+void randomize(double *X, const int rows, const int columns_X)
 {
   // Use a different seed value so that we don't get same
   // result each time we run this program
